References #162, EduSuckr side should be almost done.
NB! Please note that changes have been made to the database tables,
consult UPGRADE file for specific details. Added blog_base to 
participant, this should allow the system to work with feeds with
categories / lables depending on the generator. Added a version.txt
file to hold information on the current release version. Version bump
will take place as soon as the changes are proven to be safe and
working without glitches. Extended Complex Types to handle blog_base
for participants. Changed several methods to use blog_base instead of
blog so that all the connections would be intact. Added a check for
newly imported comments to become hidden if their parent post is that.
Newly added feature still needs some testing to prove itself working.
Please note that category / lable feeds are only allowed for course
participants, in case of course blog the base URL should be used.